refactor: rename cancel button to warning button
